Implement Azure AD omniauth provider
To set up this up, you are required to create a new "App registration"
within the Azure Active Directory service. Keep note of the Application
ID, as this will be the Client ID required for the AAD_CLIENT_ID env
variable. You will also be required to add the callback URL to the
"reply URL" within the app registry properties, e.g.
'http://localhost:3000/auth/azureactivedirectory/callback'.

Currently there is an issue with the omniauth-azure-activedirectory gem
whereby we need to require the 'net/http' gem in order for the callback
to be registered. See the following open issue for more details:
https://github.com/AzureAD/omniauth-azure-activedirectory/issues/21