Rename payload helper with __modifier type for clarity
An attempt to make it more obvious when reading the tests or anywhere
the _type_ of payload is used. The idea is that we'd have the happy path
payloads, but there will be situations where different payloads are
required to show different scenarios on the front end. I am hoping that
this method naming pattern can help show it.