[WIP] Attempt to get webpacker workign for stylesheets
I think I might be attempting more than what Webpacker was built for,
but I believe we can speed up the asset compilation time by switching
from Sprockets to Webpacker. Already, but using the
./bin/webpacker-dev-server executabe we no longer have to wait up to a
minute for the page to respond.

However, the current issue (as shown in
app/javascript/packs/styles.scss) is that the SCSS partials are not
being included.