Import work orders into the graph db
We need to import the work orders as well so that we have a coherent structure
in the graph db.

Reactors importing into a GraphModelImporter - this keeps the jobs slimmer and
allows us to re-use the code for out bulk import. Also linking the work orders
here as it's easier to do it now rather than backfilling.