Create a 'primary' path through the connected graph to improve performance
Tag the citation with extra=true if we are already part of the related
work_order structure and extra=false if we are joining for the first time.
This creates a set if citations that can be followed efficiently.

There is a chance that duplicate paths can still be created. A chain of
22 work orders (w1..w22) could have a 'extra=false' citation created if
a note links w1 and w22 since we only check 20 hops for a duplicate.