Background job to delete very old notes
Our graph DB will eventually fill up if we keep on storing all the notes.
We want to keep a large number of them though since they will let us fall back
to using the graph db as a genertal index if the UH db is under too much load.